move cli to own class
